pyproject.toml
[tool.poetry]
name = "coral-vision"
version = "0.2.0"
description = "Coral Vision - Offline Face Recognition with Edge TPU Acceleration and pgvector Support"
authors = ["Avishay Bar <avishay.il@gmail.com>"]
readme = "README.md"
packages = [{ include = "coral_vision" }]
keywords = ["face-recognition", "edge-tpu", "coral", "tensorflow-lite", "pgvector", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
pyttsx3 = "^2.90"
opencv-python = "^4.8.0"
numpy = "^1.24.0"
Pillow = "^10.0.0"
typer = "^0.12.3"
tensorflow = "~2.15"
flask = "^3.0.0"
flask-socketio = "^5.3.0"
python-socketio = "^5.10.0"
eventlet = "^0.33.0"
werkzeug = "^3.0.0"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.4"
flask-limiter = "^3.5.0"
tenacity = "^8.2.3"

# Explicitly bind pycoral to the Coral index
pycoral = { version = "~2.0", source = "coral" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"
requests = "^2.31.0"
black = "==22.6.0"
isort = "==5.11.5"
flake8 = "==4.0.1"
flake8-docstrings = "==1.6.0"
coverage = "*"
yamllint = "*"
pre-commit = "*"

[[tool.poetry.source]]
name = "coral"
url = "https://google-coral.github.io/py-repo/"
priority = "supplemental"

[tool.poetry.scripts]
coral-vision = "coral_vision.cli:app"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
